+++ features/app/src/Controller/CatchAll.php
@@ -1,18 +1,20 @@
 <?php
-
 /**
- * This file is part of WebStack
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * Created by PhpStorm.
+ * User: fsilva
+ * Date: 06-12-2017
+ * Time: 16:17
  */
-
 
 namespace Features\App\Controller;
 
-
 use Slick\WebStack\Controller;
 
+/**
+ * Catch All Controller
+ *
+ * @package Features\App\Controller
+ */
 class CatchAll extends Controller
 {