+++ spec/Controller/ContextSpec.php
@@ -1,28 +1,35 @@
 <?php
 
+/**
+ * This file is part of slick/web_stack package
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace spec\Slick\WebStack\Controller;
 
+use Aura\Router\Route;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Slick\Di\ContainerInjectionInterface;
-use Slick\Http\Uri;
+use Slick\Http\Message\Uri;
 use Slick\WebStack\Controller\Context;
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
 use Slick\WebStack\Controller\ControllerContextInterface;
 use Slick\WebStack\Service\UriGeneratorInterface;
 
+/**
+ * ContextSpec specs
+ *
+ * @package spec\Slick\WebStack\Controller
+ */
 class ContextSpec extends ObjectBehavior
 {
 
-    function let(UriGeneratorInterface $uriGenerator)
+    function let(ServerRequestInterface $request, UriGeneratorInterface $uriGenerator)
     {
-        $this->beConstructedWith($uriGenerator);
-    }
-
-    function it_is_initializable()
-    {
-        $this->shouldHaveType(Context::class);
+        $route = (new Route())->attributes(['foo' =>'bar', 'bar' => 'baz']);
+        $this->beConstructedWith($request, $route, $uriGenerator);
     }
 
     function its_a_controller_context()
@@ -30,134 +37,97 @@ class ContextSpec extends ObjectBehavior
         $this->shouldBeAnInstanceOf(ControllerContextInterface::class);
     }
 
-    function it_implements_a_container_injection_for_initialization()
+    function it_is_initializable()
     {
-        $this->shouldImplement(ContainerInjectionInterface::class);
+        $this->shouldHaveType(Context::class);
     }
 
-    function it_uses_an_http_request_and_response(
-        ServerRequestInterface $request,
-        ResponseInterface $response
+    function it_can_have_a_list_of_query_params(
+        ServerRequestInterface $request
     )
     {
-        $this->register($request, $response)->shouldBe($this->getWrappedObject());
+        $data = ['foo' =>'bar', 'bar' => 'baz'];
+        $request->getQueryParams()->willReturn($data);
+        $this->queryParam()->shouldBe($data);
     }
 
-    function it_retrieves_the_server_parameters_on_post(
-        ServerRequestInterface $request,
-        ResponseInterface $response
+    function it_can_retrieve_a_single_query_param_value(
+        ServerRequestInterface $request
     )
     {
-        $request->getServerParams()->willReturn(
-            [
-                'foo' => 'bar',
-                'bar' => 'baz'
-            ]
-        );
-        $this->register($request, $response);
-        $this->getPost()->shouldBe([
-            'foo' => 'bar',
-            'bar' => 'baz'
-        ]);
+        $data = ['foo' =>'bar', 'bar' => 'baz'];
+        $request->getQueryParams()->willReturn($data);
+        $this->queryParam('foo')->shouldBe('bar');
     }
 
-    function it_can_retrieve_a_single_named_parameter(
-        ServerRequestInterface $request,
-        ResponseInterface $response
+    function it_can_have_a_list_of_post_parameters(
+        ServerRequestInterface $request
     )
     {
-        $request->getServerParams()->willReturn(
-            [
-                'foo' => 'bar',
-                'bar' => 'baz'
-            ]
-        );
-        $this->register($request, $response);
-        $this->getPost('bar')->shouldBe('baz');
+        $data = ['foo' =>'bar', 'bar' => 'baz'];
+        $request->getParsedBody()->willReturn($data);
+        $this->postParam()->shouldBe($data);
     }
 
-    function it_returns_a_default_value_if_param_is_not_set(
-        ServerRequestInterface $request,
-        ResponseInterface $response
+    function it_will_return_a_default_value_for_missing_parameters(
+        ServerRequestInterface $request
     )
     {
-        $request->getServerParams()->willReturn(
-            [
-                'foo' => 'bar',
-                'bar' => 'baz'
-            ]
-        );
-        $this->register($request, $response);
-        $this->getPost('boo', 'test')->shouldBe('test');
+        $data = ['foo' =>'bar', 'bar' => 'baz'];
+        $request->getParsedBody()->willReturn($data);
+        $this->postParam('baz', 'foo')->shouldBe('foo');
     }
 
-    function it_can_get_URL_parameters(
-        ServerRequestInterface $request,
-        ResponseInterface $response
-    )
+    function it_can_retrieve_a_route_parameter()
     {
-        $request->getQueryParams()->willReturn(
-            [
-                'foo' => 'bar',
-                'bar' => 'baz'
-            ]
-        );
-        $this->register($request, $response);
-        $this->getQuery('boo', 'test')->shouldBe('test');
-        $this->getQuery('bar')->shouldBe('baz');
-        $this->getQuery()->shouldBe([
-            'foo' => 'bar',
-            'bar' => 'baz'
-        ]);
+        $this->routeParam('foo')->shouldBe('bar');
     }
 
-    function it_can_set_a_redirecting_http_response(
-        ServerRequestInterface $request,
-        ResponseInterface $response,
-        UriGeneratorInterface $uriGenerator
+    function it_has_a_server_request(ServerRequestInterface $request)
+    {
+        $this->request()->shouldBe($request);
+    }
+
+    function it_can_check_if_current_request_has_a_given_method(
+        ServerRequestInterface $request
     )
     {
-        $this->beConstructedWith($uriGenerator);
+        $request->getMethod()->willReturn('PUT');
+        $this->requestIs('put')->shouldBe(true);
+    }
 
-        $location = 'home';
-        $response->withStatus(302)->willReturn($response);
-        $response->withHeader('location', '/pages/index')
+    function it_can_be_set_with_a_response(ResponseInterface $response)
+    {
+        $this->setResponse($response)->shouldBe($this->getWrappedObject());
+        $this->response()->shouldBe($response);
+    }
+
+    function it_can_set_a_redirect_response(UriGeneratorInterface $uriGenerator)
+    {
+        $uriGenerator->generate('/some/page', ['foo' => 'bar'])
             ->shouldBeCalled()
-            ->willReturn($response);
-        $uriGenerator->setRequest($request)
-            ->shouldBeCalled();
-        $uriGenerator->generate($location, [])
-            ->shouldBeCalled()
-            ->willReturn(new Uri('/pages/index'));
+            ->willReturn(new Uri('http://example.com/some/page?foo=bar'));
 
-        $this->register($request, $response);
-        $this->redirect($location);
-
-        $response->withStatus(302)->shouldHaveBeenCalled();
+        $this->redirect('/some/page', ['foo' => 'bar']);
+        $response = $this->response();
+        $response->shouldBeAnInstanceOf(ResponseInterface::class);
+        $response->getStatusCode()->shouldBe(302);
+        $response->getHeaderLine('Location')->shouldBe('http://example.com/some/page?foo=bar');
+        $this->handlesResponse()->shouldBe(true);
     }
 
-    function it_can_set_the_disable_rendering_flag_on_request(
-        ServerRequestInterface $request,
-        ResponseInterface $response
-    )
+    function it_can_set_the_template_to_be_used(ServerRequestInterface $request)
     {
-        $request->withAttribute('rendering', false)->willReturn($request);
-        $this->register($request, $response);
-        $this->disableRendering()->shouldBe($this->getWrappedObject());
-        $request->withAttribute('rendering', false)->shouldHaveBeenCalled();
-    }
-
-    function it_can_set_the_rendering_view(
-        ServerRequestInterface $request,
-        ResponseInterface $response
-    )
-    {
-        $this->register($request, $response);
-
-        $request->withAttribute('view', 'path/to/view.twig')
+        $request->withAttribute('template', 'some/path')
             ->shouldBeCalled()
             ->willReturn($request);
-        $this->setView('path/to/view.twig')->shouldReturn($this->getWrappedObject());
+        $this->useTemplate('some/path')->shouldBe($this->getWrappedObject());
     }
 
+    function it_can_change_the_request(ServerRequestInterface $otherRequest, ServerRequestInterface $request)
+    {
+        $this->changeRequest($otherRequest)->shouldBe($this->getWrappedObject());
+        $this->request()->shouldBe($otherRequest);
+        $this->request()->shouldNotBe($request);
+    }
 }